% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_mr_aic.R
\name{bi_mr_aic}
\alias{bi_mr_aic}
\title{comparing bi-direction causal models using AIC}
\usage{
bi_mr_aic(X.set, Y.set)
}
\arguments{
\item{X.set}{Trait X ascertainment data set}

\item{Y.set}{Trait Y ascertainment data set}
}
\value{
relative likelihood of the best non-causal model compared to the best causal model.
}
\description{
calculate the Akaike information criterion (AIC) for four causal inference models;
implementation of the Pickrell's method
}
