% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_mr_aic.R
\name{lh_M}
\alias{lh_M}
\title{approximate likelyhood function}
\usage{
lh_M(x_hat, x, n_x, y_hat, y, n_y)
}
\arguments{
\item{x_hat}{Fisher z-transformation of trait X data set}

\item{x}{causal coefficients of trait X data set}

\item{n_x}{length of trait X data set}

\item{y_hat}{Fisher z-transformation of trait Y data set}

\item{y}{causal coefficients of trait Y data set}

\item{n_y}{length of trait Y data set}
}
\value{
likelyhood
}
\description{
approximate likelihood for the two correlation coefficients
}
