% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_ivw.R
\name{mr_ivw}
\alias{mr_ivw}
\title{Inverse-variance weighted method for MR using the input data set}
\usage{
mr_ivw(data.set)
}
\arguments{
\item{data.set}{the input summary statistic data set}
}
\value{
a list with beta, se and p
}
\description{
Calculating the causal inference estimate(beta), se and p value using IVW
}
