% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{rd_info}
\alias{rd_info}
\title{Subset the file information}
\usage{
rd_info(file_to_rd)
}
\arguments{
\item{file_to_rd}{the summary statistics file}
}
\value{
a list contains \code{beta}, \code{se}, \code{n};
or the \code{expo.beta}, \code{expo.se}, \code{expo.n} of exposure trait in exposure ascertainment data set
the \code{out.beta}, \code{out.se}, \code{out.n} of outcome trait in exposure ascertainment data set
}
\description{
Subset the file to get summary statistics information
}
