% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_mr_aic.R
\name{z_trans}
\alias{z_trans}
\title{Fisher’s Z-transformation}
\usage{
z_trans(x, y)
}
\arguments{
\item{y}{vectors of effect sizes on Traint Y}
}
\value{
A list with  correlation coefficient \code{rho} and  Fisher z-transformation \code{z}
}
\description{
Compute the Fisher’s Z-transformation of each data set
}
\details{
x vectors of effect sizes on Traint X
}
